+++ test/services/jobs/return-logs/create-current-return-cycle.service.test.js
@@ -15,12 +15,12 @@ const {
 } = require('../../../../app/lib/return-cycle-dates.lib.js')
 
 // Thing under test
-const GenerateReturnCycleService = require('../../../../app/services/jobs/return-logs/generate-return-cycle.service.js')
+const GenerateCurrentReturnCycleService = require('../../../../app/services/jobs/return-logs/generate-current-return-cycle.service.js')
 
 describe('Generate return cycle service', () => {
   describe('when summer is true', () => {
     it('should create a summer return cycle with the correct values and return the id', async () => {
-      const result = await GenerateReturnCycleService.go(true)
+      const result = await GenerateCurrentReturnCycleService.go(true)
 
       expect(result.startDate).to.equal(cycleStartDateAsISO(true))
       expect(result.endDate).to.equal(cycleEndDateAsISO(true))
@@ -32,7 +32,7 @@ describe('Generate return cycle service', () => {
 
   describe('when summer is false', () => {
     it('should create an all year return cycle with the correct values and return the id', async () => {
-      const result = await GenerateReturnCycleService.go(false)
+      const result = await GenerateCurrentReturnCycleService.go(false)
 
       expect(result.startDate).to.equal(cycleStartDateAsISO(false))
       expect(result.endDate).to.equal(cycleEndDateAsISO(false))
